bump the coverage
